gui 70% cleaned
