+++ pkg/gui/gui_init.go
@@ -16,7 +16,7 @@ var (
 func GuiInit() {
 	var guiInitBenchmark *core.BenchmarkTimer
 	if config.DEBUGGING {
-		guiInitBenchmark = core.StartBenchmark("GuiInit()", true)
+		guiInitBenchmark = core.StartBenchmark("gui.GuiInit()", true)
 	}
 	// Initialize GTK.
 	gtk.Init(nil)
